ai-kb-rag-pytorch/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
transformers==4.44.2
torch==2.3.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.5.1
tqdm==4.66.4
pydantic==2.9.2
redis==5.0.8
